requirements.txt
asgiref==3.8.1
attrs==25.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
dj-config-url==0.1.1
dj-database-url==2.3.0
Django==5.1.7
django-channels==0.7.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
oauthlib==3.2.2
packaging==24.2
Pillow
psycopg==3.2.5
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
